+++ src/components/combo/controllers/navigation.ts
@@ -151,6 +151,11 @@ export class ComboNavigationController<T extends object> {
         // before the Shift+Tab behavior kicks in.
         this._host.focus();
       }
+
+      if (this._host.singleSelect && this.active > -1) {
+        this._config.interactions.select(this.active);
+      }
+
       await this._config.interactions.hide();
     }
   };