+++ src/index.ts
@@ -108,15 +108,19 @@ export {
 export {
   type ICalendarResourceStrings,
   type ICarouselResourceStrings,
+  type IChatResourceStrings,
   type IChipResourceStrings,
   type IComboResourceStrings,
+  type IDatePickerResourceStrings,
   type IDateRangePickerResourceStrings,
   type IFileInputResourceStrings,
   type ITreeResourceStrings,
   CalendarResourceStringsEN,
   CarouselResourceStringsEN,
+  ChatResourceStringsEN,
   ChipResourceStringsEN,
   ComboResourceStringsEN,
+  DatePickerResourceStringsEN,
   DateRangePickerResourceStringsEN,
   FileInputResourceStringsEN,
   TreeResourceStringsEN,
@@ -144,6 +148,7 @@ export type { IgcDropdownComponentEventMap } from './components/dropdown/dropdow
 export type { IgcExpansionPanelComponentEventMap } from './components/expansion-panel/expansion-panel.js';
 export type { IgcInputComponentEventMap } from './components/input/input-base.js';
 export type { IgcInputComponentEventMap as IgcMaskInputComponentEventMap } from './components/input/input-base.js';
+export type { IgcNavDrawerComponentEventMap } from './components/nav-drawer/nav-drawer.js';
 export type { IgcFileInputComponentEventMap } from './components/file-input/file-input.js';
 export type { IgcRadioComponentEventMap } from './components/radio/radio.js';
 export type { IgcRatingComponentEventMap } from './components/rating/rating.js';