+++ stories/input.stories.ts
@@ -32,9 +32,8 @@ const metadata: Meta<IgcInputComponent> = {
   },
   argTypes: {
     value: {
-      type: 'string | Date',
+      type: 'string',
       description: 'The value of the control.',
-      options: ['string', 'Date'],
       control: 'text',
     },
     type: {
@@ -161,7 +160,7 @@ export default metadata;
 
 interface IgcInputArgs {
   /** The value of the control. */
-  value: string | Date;
+  value: string;
   /** The type attribute of the control. */
   type: 'text' | 'email' | 'number' | 'password' | 'search' | 'tel' | 'url';
   /** Makes the control a readonly field. */